# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "App Status API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/health": {
            "get": {
                "description": "Query the health of the service",
                "summary": "Health",
                "operationId": "health",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/model.EmptyBody"
                        }
                    },
                    "500": {
                        "description": "Server is down",
                        "schema": {
                            "$ref": "#/definitions/model.EmptyBody"
                        }
                    }
                }
            }
        },
        "/status/": {
            "post": {
                "description": "Publish a new app status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Publish App status",
                "operationId": "publish-app-status",
                "parameters": [
                    {
                        "description": "New App Status",
                        "name": "status-request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ReleaseRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {},
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/status/version/{version}/{platform}": {
            "get": {
                "description": "Query app status for a given app release version",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get App Status",
                "operationId": "get-app-status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "app version",
                        "name": "version",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "App platform IOS, Android",
                        "name": "platform",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/model.ReleaseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/model.ReleaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.EmptyBody": {
            "type": "object"
        },
        "model.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "model.ReleaseRequest": {
            "type": "object",
            "required": [
                "platform",
                "version"
            ],
            "properties": {
                "platform": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "model.ReleaseResponse": {
            "type": "object",
            "properties": {
                "flags": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "boolean"
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}